# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minecraft Hosting Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all servers
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a new server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerRequest"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a server by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "404":
          description: Server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/power:
    post:
      operationId: serverPowerAction
      tags: [servers]
      summary: Start, stop, or restart a server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PowerActionRequest"
      responses:
        "200":
          description: Action performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/command:
    post:
      operationId: sendServerCommand
      tags: [servers]
      summary: Send a command to the server console
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandRequest"
      responses:
        "200":
          description: Command sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/stats:
    get:
      operationId: getServerStats
      tags: [servers]
      summary: Get server resource stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStats"
  /servers/{id}/logs:
    get:
      operationId: getServerLogs
      tags: [servers]
      summary: Get recent console logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lines
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Console logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
...
  /servers/{id}/update:
    patch:
      operationId: updateServer
      tags: [servers]
      summary: Update server settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerRequest"
      responses:
        "200":
          description: Server updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
...
  /servers/{id}/subdomains:
    get:
      operationId: listServerSubdomains
      tags: [servers]
      summary: List subdomains for a server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subdomains list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subdomain"
    post:
      operationId: createServerSubdomain
      tags: [servers]
      summary: Create a subdomain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubdomainRequest"
      responses:
        "201":
          description: Subdomain created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
  /servers/{id}/subdomains/{subId}:
    delete:
      operationId: deleteServerSubdomain
      tags: [servers]
      summary: Delete a subdomain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: subId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subdomain deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Server:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [starting, running, stopping, stopped, crashed]
        version:
          type: string
        port:
          type: number
        memory:
          type: number
        disk:
          type: number
        nodeId:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        playerCount:
          type: number
        maxPlayers:
          type: number
        javaVersion:
          type: string
        serverType:
          type: string
          enum: [vanilla, paper, spigot, forge, fabric, bungeecord]
      required:
        - id
        - name
        - status
        - version
        - port
        - memory
        - disk
        - nodeId
        - userId
        - createdAt
        - updatedAt
        - serverType
    CreateServerRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        version:
          type: string
        port:
          type: number
        memory:
          type: number
        disk:
          type: number
        nodeId:
          type: string
        userId:
          type: string
        maxPlayers:
          type: number
        javaVersion:
          type: string
        serverType:
          type: string
          enum: [vanilla, paper, spigot, forge, fabric, bungeecord]
      required:
        - name
        - version
        - port
        - memory
        - disk
        - nodeId
        - userId
        - serverType
    PowerActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [start, stop, restart, kill]
      required:
        - action
    CommandRequest:
      type: object
      properties:
        command:
          type: string
      required:
        - command
    ServerStats:
      type: object
      properties:
        serverId:
          type: string
        cpuPercent:
          type: number
        memoryUsed:
          type: number
        memoryLimit:
          type: number
        diskUsed:
          type: number
        diskLimit:
          type: number
        networkRx:
          type: number
        networkTx:
          type: number
        uptime:
          type: number
        playerCount:
          type: number
      required:
        - serverId
        - cpuPercent
        - memoryUsed
        - memoryLimit
        - diskUsed
        - diskLimit
        - networkRx
        - networkTx
        - uptime
        - playerCount
    LogsResponse:
      type: object
      properties:
        serverId:
          type: string
        logs:
          type: array
          items:
            type: string
      required:
        - serverId
        - logs
...
    UpdateServerRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        memory:
          type: number
        disk:
          type: number
        maxPlayers:
          type: number
        port:
          type: number
...
    Subdomain:
      type: object
      properties:
        id:
          type: string
        serverId:
          type: string
        subdomain:
          type: string
        targetPort:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - subdomain
        - targetPort
        - createdAt
    CreateSubdomainRequest:
      type: object
      properties:
        subdomain:
          type: string
      required:
        - subdomain