# lib/api-zod/src/generated/types/createSubdomainRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSubdomainRequest {
  subdomain: string;
}