# lib/api-zod/src/generated/types/subdomain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Subdomain {
  id: string;
  serverId: string;
  subdomain: string;
  targetPort: string;
  createdAt: Date;
}