# lib/api-zod/src/generated/types/updateServerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServerRequest {
  name?: string;
  description?: string;
  memory?: number;
  disk?: number;
  maxPlayers?: number;
  port?: number;
}